requirements.txt
streamlit==1.27.1
pillow==10.0.1